backend/requirements.txt
asgiref==3.6.0
astroid==2.15.2
attrs==22.2.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-material-admin==1.8.6
django-oauth-toolkit==2.2.0
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jwcrypto==1.4.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.2
pytest==7.3.0
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.6
tzdata==2022.7
urllib3==1.26.14
wrapt==1.15.0
